better orgainzed tests
